refactor: move element styles to separate model
